Extract autoloading process
The process sits in its own file now to ease future improvements.
